Remove Qodana configuration and workflow
Deleted Qodana configuration file and associated GitHub workflow. Qodana was initially set up for code quality analysis, but due to certain restrictions, performance inefficiency, and conflicts with other parts of the existing system, it has been decided to remove Qodana from the project. Approaches will be reconsidered for code quality checks and analysis in the future.