Replace pgx with sql package in dbService
The pgx dependency has been removed in dbService.go and replaced with standard sql package. This was done to align error handling with the standard sql package, specifically for checking 'No rows in result set' error. The undesired use of pgx introduced an inconsistency with the rest of the code which predominantly uses the built-in SQL package for managing database interactions.