Reduce memory allocations in metrics collection
Removed unnecessary 'return' statement in 'collectAdvancedMetrics'. Also, reduced the capacity of 'm' slice from 20 to 15 in this method, as it is sufficient for our requirements and helps to decrease memory usage.

This change improves the performance of the metrics agent by reducing unnecessary memory allocations. This is part of ongoing optimizations aiming to make the agent less resource-intensive.