Move metrics initialization inside loops
In this commit, the initialization of 'metrics' slice has been moved inside the loops in both 'CollectRuntime' and 'CollectAdvanced' methods of 'MetricService'. This change ensures that 'metrics' is freshly initialized for each iteration of the loop, thus avoiding potential accumulation or overwriting of data from preceding iterations. This improves data accuracy and reliability of metrics collection.