Upgrade Go version to 1.21 and release-downloader to v1.8 in Github Actions
The Go version used in Github workflows namely `statictest.yml` and `devopstest.yml` has been upgraded from 1.19 to 1.21. Additionally, the version of GitHub Actions third-party utility 'release-downloader' has been updated to v1.8 from v1 and v1.2 respectively. This upgrade is required to leverage the latest features and improvements in the Go language and the utility tool for more efficient build and testing process.