Add logging for returned metrics in metric.go
A line of code has been added to the "metric.go" file to log the JSON response of metrics. By logging the response, we will now have visibility into the data returned by this function which can assist us with better debugging and understanding the information flow in our application.