Remove unnecessary logs and simplify hash updates
Enhanced the maintainability and efficiency of the metrics services by reducing unnecessary logging in metricService.go and metric.go for cleaner console output. In dbService.go, removed a redundant hash existence check for assumption of pre-existing hash, thereby making the condition for hash generation simpler and easier to read. An added lock and a temporary key for delta calculation were introduced for further tests.