Remove logging from hash function and streamline code
This commit removes a line of logging from the 'Hash' function in 'hash.go'. This logging was unnecessary and could potentially add overhead to the hash calculation, especially for high traffic applications. Additionally, it removes the setting of 'del' and 'key' values as they were not being used, simplifying the code and improving readability. This commit also removes the unneeded hash existence check, optimising the hash calculation process.