Simplify error handling and enhance response writing in metric.go
Removed checking for an error while writing the error response. This eliminates one layer of error handling and simplifies the structure of the code. Also replaced usage of `Write` method with `Fprint` for returning metric data in json format, thus enhancing readability. These modifications improve maintainability and clarity while ensuring that the correct response is returned in case of errors.