Add Swagger documentation to Metrics Alerting Service API
This commit adds Swagger documentation to the Metrics Alerting Service API. This ensures that the API functionality is described and accessible via a user-friendly interface. Both documentation files (JSON+Yaml) and auto-generated 'docs.go' by 'swaggo' are added. The defined routes are also decorated with the Swagger annotations for a more detailed description. This commit also adds error handling for response writing. If an error occurs while writing the response, it will now be logged for debugging purposes. Apart from that, some changes may be seen in other files due to the auto formatter's effect.